A new approach to the Gemfile which works with OBS and Docker 'shared volume' development.
Clean up the Gemfile to use git for the areas we typically develop
with.

Allow an environment variable (USE_GIT) to drive using git versions of
ruote/ruote-kit/amqp etc to allow config local.* to override PATH in
Bundler. This approach works with OBS tar_git builds too.

Allow boss and associated gems to to built as a gem and installed from
a local file:// based gem repository (./gemrepo/) using the
./dev/build_gems script.

Signed-off-by: David Greaves <david.greaves@jolla.com>